+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @orderly.network/eslint-config
 
+## 2.8.12
+
+### Patch Changes
+
+- publish
+- 117b23a: publish
+
+## 2.8.12-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 2.8.11
+
+### Patch Changes
+
+- 1ed7590: publish
+- publish
+- e7fb6c1: publish
+
+## 2.8.11-alpha.1
+
+### Patch Changes
+
+- publish
+
 ## 2.8.11-alpha.0
 
 ### Patch Changes